pyproject.toml
[project]
name = "pulsestream"
version = "0.1.0"
description = "Real-time market intel and news radar"
authors = [
    {name = "Achyuth", email = "achyuthrayal2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2,<0.136.0",
    "uvicorn>=0.42.0,<0.43.0",
    "sqlalchemy>=2.0.48,<3.0.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "requests>=2.33.1,<3.0.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "gnews>=0.4.3,<0.5.0",
    "newspaper3k>=0.2.8,<0.3.0",
    "groq>=1.1.2,<2.0.0",
    "lxml-html-clean>=0.4.4,<0.5.0",
    "apscheduler>=3.11.2,<4.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"